Fixing several small bugs: - Unreferenced formal parameter in allocator allocation function is now referenced and has a valid value set - Adding "Location" value to debug layer output - Attempting to fix issues with dangling pointers to vertex attribute and binding descriptions in mesh class - Setting an actual Max LOD level in the terrain quadtree class - Setting the LOD switch ratio in the terrain quadtree class - Remove unreferenced variable in NodeRenderer - Fixing error that disabled validation layers when creating a logical device. - Type casting warnings suppressed wherever found - Removed FastNoise, deprecated/unused atmosphere tests. - Removed unreferenced swapchain parameter from most scenes and complex objects. Was used before dynamic state used to set viewport+scissor rect - Fixed using class member name (size) in several methods for the Buffer class and for a method in the GridMesh class - Fixed two errors where not all control paths for the allocation subsystem's methods returned a value - Removed LodePNG and saving to PNG file from NodeRenderer system - Fixed several instances where a VkResult instance was initialized, but not checked with VkAssert()

+++ DiamondDogs/engine/terrain/DataProducer.cpp
@@ -36,7 +36,7 @@ namespace vulpes {
 
 	DataProducer::DataProducer(const Device * _parent) : parent(_parent) {
 		
-		pipelineCache = std::make_unique<PipelineCache>(parent, typeid(DataProducer).hash_code() + numProducers);
+		pipelineCache = std::make_unique<PipelineCache>(parent, static_cast<int16_t>(typeid(DataProducer).hash_code() + numProducers));
 
 		// we offset the hash used for pipeline cache by this: that way,
 		// each producer instance gets its own cache (so, split instances among distinct task types to leverage cache)
@@ -102,7 +102,7 @@ namespace vulpes {
 		pool_info.queueFamilyIndex = parent->QueueFamilyIndices.Compute;
 
 		computePool = std::make_unique<CommandPool>(parent, pool_info, true);
-		computePool->AllocateCmdBuffers(availQueues.size());
+		computePool->AllocateCmdBuffers(static_cast<uint32_t>(availQueues.size()));
 
 		spareQueue = parent->GeneralQueue(0);
 		pool_info.queueFamilyIndex = parent->QueueFamilyIndices.Graphics;
@@ -324,7 +324,7 @@ namespace vulpes {
 
 		vkCmdBindPipeline(cmd_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipelines[curr_idx]);
 		vkCmdBindDescriptorSets(cmd_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipelineLayout, 0, 1, &descriptorSet, 0, nullptr);
-		vkCmdDispatch(cmd_buffer, ceil(request->Width / 16), ceil(request->Height / 16), 1);
+		vkCmdDispatch(cmd_buffer, static_cast<uint32_t>(ceil(request->Width / 16)), static_cast<uint32_t>(ceil(request->Height / 16)), 1);
 		vkCmdPipelineBarrier(cmd_buffer, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT, 0, 0, nullptr, 1, &computeCompleteBarrier, 0, nullptr);
 		vkCmdCopyBuffer(cmd_buffer, request->Output->vkHandle(), request->Result->vkHandle(), 1, &output_to_result_copy);
 		vkCmdPipelineBarrier(cmd_buffer, VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_HOST_BIT, 0, 0, nullptr, 1, &hostTransitionBarrier, 0, nullptr);
